# src/openapi.yaml
openapi: 3.0.3
info:
  title: ScoutOff API
  description: >
    Backend API for ScoutOff — a decentralised football scouting platform built on the Stellar
    blockchain (Soroban smart contracts).
  version: 1.0.0
  contact:
    name: ScoutOff Engineering
servers:
  - url: /api
    description: Stable API (version-less alias)
  - url: /api/v1
    description: API v1
  - url: /api/v2
    description: API v2 (includes SSE stream endpoint)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >-
        Long-lived scout API key (alternative to a Bearer JWT). Restricted keys may only perform
        operations covered by their granted scopes (#1019). Legacy keys without a scope list are
        unrestricted.
...
paths:
...
  /auth/challenge:
    get:
      tags:
        - auth
      summary: Issue a SEP-10-style challenge transaction XDR for the given Stellar account.
      operationId: getChallenge
      description: |-
        Issue a SEP-10-style challenge transaction XDR for the given Stellar
        account. The client signs it and exchanges it for a token via POST
        /auth/token.
      parameters:
        - name: account
          in: query
          required: false
          schema:
            type: string
          description: Stellar public key (G...) to build the challenge for
      responses:
        '200':
          description: '{ challenge: string, networkPassphrase: string }'
        '400':
          description: '{ success: false, error: string } - Missing/invalid account'
      security: []
  /auth/logout:
    post:
      tags:
        - auth
      summary: >-
        Revoke the caller's access token and any associated refresh token (identified by the `jti`
        in the bearer token).
      operationId: postLogout
      description: |-
        Revoke the caller's access token and any associated refresh token
        (identified by the `jti` in the bearer token).
      requestBody:
        required: true
        description: '{ refreshToken?: string }'
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  minLength: 1
              additionalProperties: false
      responses:
        '200':
          description: '{ success: true }'
        '401':
          description: — missing/invalid bearer token
      security:
        - bearerAuth: []
  /auth/refresh:
    post:
      tags:
        - auth
      summary: >-
        Exchange a valid refresh token for a new access token + refresh token pair (refresh token
        rotation).
      operationId: postRefresh
      description: |-
        Exchange a valid refresh token for a new access token + refresh token pair
        (refresh token rotation). The old refresh token is revoked immediately.
      requestBody:
        required: true
        description: '{ refreshToken: string }'
        content:
          application/json:
            schema:
              type: object
              properties:
                refreshToken:
                  type: string
                  minLength: 1
              required:
                - refreshToken
              additionalProperties: false
      responses:
        '200':
          description: '{ accessToken, refreshToken, expiresAt }'
        '401':
          description: '{ success: false, error } — invalid, expired, or revoked refresh token'
      security: []
  /auth/token:
    post:
      tags:
        - auth
      summary: Exchange a signed challenge transaction for an access + refresh token pair.
      operationId: postToken
      description: |-
        Exchange a signed challenge transaction for an access + refresh token
        pair. The caller's role is derived from the verified account: configured
        admin wallets always get `admin`, otherwise the requested `role` (default
        `player`) is used.
      requestBody:
        required: true
        description: '{ transaction: string, role?: ''validator'' | ''player'' | ''scout'' } - Signed challenge XDR'
        content:
          application/json:
            schema:
              type: object
              properties:
                transaction:
                  type: string
                  minLength: 1
                role:
                  type: string
                  enum:
                    - player
                    - scout
                    - validator
                    - admin
              required:
                - transaction
              additionalProperties: false
      responses:
        '200':
          description: '{ token, accessToken, refreshToken, account, expiresAt }'
        '400':
          description: '{ success: false, error: string } - Invalid body or malformed XDR'
        '401':
          description: '{ success: false, error: string } - Invalid signature or expired challenge'
      security: []
...